Baidu POI Search Demo and Marker Refacter
Added Baidu POI Search, Demo implementation.
Change AR Marker class, a BasicMarker class to provider interface.
The GoogleMarker and BaiduMarker provider two different implementations.
GlobalARData also add fields for gaining location data from Google and
Baidu API at the same time.

+++ src/com/mengdd/map/baidu/BaiduMapViewModel.java
@@ -127,7 +127,7 @@ public class BaiduMapViewModel extends BasicMapViewModel
 	{
 		GeoPoint centerpt = mMapView.getMapCenter();
 		int maxLevel = mMapView.getMaxZoomLevel();
-		int zoomlevel = mMapView.getZoomLevel();
+		float zoomlevel = mMapView.getZoomLevel();
 		boolean isTraffic = mMapView.isTraffic();
 		boolean isSatillite = mMapView.isSatellite();
 		boolean isDoubleClick = mMapView.isDoubleClickZooming();
